+++ CHANGELOG.md
@@ -10,8 +10,8 @@
 - Support varargs printing (example: `pd arg1, arg2, arg3`)
 - Display `run_at` run number in printout
 - Support `require 'pd`' as a shorter alternative to `require 'puts_debuggerer'`
-- Support `return: false` option to return printed String instead of returning printed object
 - Support `printer` as a Logger object or Logging::Logger (from "logging" gem). Basically any object that responds to :debug method.
+- Support `printer: false` option to return rendered String instead of printing and returning object
 
 ## 0.8.2